FEATURE: Implement CreateGalaxy component and functionality, reformat forms
