+++ client/routes.js
@@ -12,6 +12,7 @@ import OrderConfirmation from "./components/OrderConfirmation";
 import UserDashboard from "./components/UserDashboard";
 import ProductDashboard from "./components/ProductDashboard";
 import EditGalaxy from "./components/EditGalaxy";
+import CreateGalaxy from "./components/CreateGalaxy"
 
 /**
  * COMPONENT
@@ -33,6 +34,7 @@ class Routes extends Component {
           <Route exact path="/users" component={UserDashboard} />
           <Route exact path="/products" component={ProductDashboard} />
           <Route exact path="/galaxies/:galaxyId/edit" component={EditGalaxy} />
+          <Route exact path="/products/add" component={CreateGalaxy} />
           <Route
             path="/cart"
             render={(props) => (
